Assistant checkpoint: Update feedback data structure to match OpenAI response
Assistant generated file changes:
- server/routes.ts: Fix feedback data structure

---

User prompt:

We know the image is being analyzed by the model gpt-4-turbo. We know there is a problem writing the data and analysis to the database. Where in the code does this happen?  Would changing the title of the item in the database have an effect?

Please take a moment to look at this from a different perspective. Track down where there is a disconnect in the code. This should be hard, but it is, we haven't made any progress is several hours. We just keep going in circles!